Bugfixes - Add Hallowed Embers, Giant Slayer - Attempt to fix alchemist's potion

+++ src/me/neoblade298/neorogue/equipment/abilities/Chokehold.java
@@ -89,6 +89,6 @@ public class Chokehold extends Equipment {
 	public void setupItem() {
 		item = createItem(Material.COPPER_GRATE,
 				"On cast, drop a zone that lasts <white>15s</white> which buffs damage of " + GlossaryTag.TRAP.tagPlural(this) + " within the zone by "
-				+ DescUtil.yellow(damage * 100 + "%") + ".");
+				+ DescUtil.yellow((int) (damage * 100) + "%") + ".");
 	}
 }
